Add Raspberry Pi Zero 2W (BCM2837) bare-metal support for arm64
Add EL2 exception level support to the arm64 package, enabling
operation on platforms where firmware starts at EL2 (e.g., Raspberry Pi).
Previously the arm64 package was hardcoded for EL3 only (i.MX8MP).
All assembly functions (MMU, cache, exceptions, IRQ, FP) now detect
CurrentEL at runtime and use appropriate EL2 or EL3 system registers.

Split soc/bcm2835 package into architecture-specific files to support
both arm and arm64 targets. The BCM2835 peripheral IP is shared across
BCM2835/BCM2836/BCM2837 SoCs, so the same driver code works for
Pi Zero 2W (BCM2837, arm64) and earlier Pi models (arm).

Key changes:
- arm64: EL2 support in init.s, mmu.s, mmu.go, cache.s, exception.s,
  irq.s, fp.s (non-breaking: EL3 continues to work for i.MX8MP)
- soc/bcm2835: split into shared + _arm.go/_arm64.go arch files
- soc/bcm2835: local busyloop abstraction replacing arm.Busyloop
- soc/bcm2835: arch-specific timer and memory configuration
- board/raspberrypi/pizero2w: new board package for Pi Zero 2W
  (BCM2837, peripheral base 0x3F000000, 512MB RAM, GPIO 29 LED)

+++ soc/bcm2835/mem_arm64.go
@@ -6,7 +6,7 @@
 // Use of this source code is governed by the license
 // that can be found in the LICENSE file.
 
-//go:build !linkramstart && arm
+//go:build !linkramstart && arm64
 
 package bcm2835
 
@@ -15,4 +15,4 @@ import (
 )
 
 //go:linkname ramStart runtime/goos.RamStart
-var ramStart uint32 = 0x00100000
+var ramStart uint64 = 0x00100000
